refactored files to respective folders
