+++ android/src/main/AndroidManifest.xml
@@ -9,22 +9,28 @@
     <uses-permission android:name="android.permission.RECEIVE_BOOT_COMPLETED" />
 
     <application>
+
         <receiver
-            android:name=".GeofenceBroadcastReceiver"
+            android:name=".receivers.GeofenceBroadcastReceiver"
             android:enabled="true"
             android:exported="true" />
-        <receiver android:enabled="true" android:name=".OnDeviceRebootBroadcastReceiver" android:permission="android.permission.RECEIVE_BOOT_COMPLETED">
+
+        <receiver android:enabled="true"
+            android:name=".receivers.OnDeviceRebootBroadcastReceiver"
+            android:permission="android.permission.RECEIVE_BOOT_COMPLETED">
             <intent-filter>
                 <action android:name="android.intent.action.BOOT_COMPLETED" />
                 <category android:name="android.intent.category.DEFAULT" />
             </intent-filter>
         </receiver>
+
         <service
-            android:name=".GeofenceTransitionsJobIntentService"
+            android:name=".services.GeofenceTransitionsJobIntentService"
             android:exported="true"
             android:enabled="true"
             android:permission="android.permission.BIND_JOB_SERVICE" />
-        <service android:name=".OnGeoFenceEventJavaScriptTaskService" />
+        <service android:name=".services.OnGeoFenceEventJavaScriptTaskService" />
+
     </application>
 
 </manifest>
