+++ android/src/main/java/me/kiano/receivers/GeofenceBroadcastReceiver.java
@@ -1,9 +1,11 @@
-package me.kiano;
+package me.kiano.receivers;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 import android.util.Log;
+
+import me.kiano.services.GeofenceTransitionsJobIntentService;
 
 public class GeofenceBroadcastReceiver extends BroadcastReceiver {
     public static final String TAG = "RNBackgroundGeofencing";
